Renamed `image::sample` to `image::sample_with`
