+++ crates/spirv-std/src/image/sample_with.rs
