+++ tests/ui/image/image_with.rs
@@ -1,7 +1,7 @@
 // build-pass
 
 use spirv_std::spirv;
-use spirv_std::{arch, image::sample, image::ImageWithMethods, Image, Sampler};
+use spirv_std::{arch, image::sample_with, image::ImageWithMethods, Image, Sampler};
 
 #[spirv(fragment)]
 pub fn main(
@@ -10,8 +10,8 @@ pub fn main(
     #[spirv(descriptor_set = 0, binding = 2)] image2: &Image!(2D, type=f32, sampled),
     output: &mut glam::Vec4,
 ) {
-    let t1 = image1.fetch_with(glam::IVec2::new(0, 0), sample::sample_index(1));
-    let t2 = image2.sample_with(*sampler, glam::Vec2::new(0.5, 0.5), sample::bias(1.0));
-    let t3 = image2.sample_with(*sampler, glam::Vec2::new(0.5, 0.5), sample::lod(2.0));
+    let t1 = image1.fetch_with(glam::IVec2::new(0, 0), sample_with::sample_index(1));
+    let t2 = image2.sample_with(*sampler, glam::Vec2::new(0.5, 0.5), sample_with::bias(1.0));
+    let t3 = image2.sample_with(*sampler, glam::Vec2::new(0.5, 0.5), sample_with::lod(2.0));
     *output = t1 + t2 + t3;
 }
